refactor tests and security configuration, fix delete function, all tests passed, rhum crud tested on postman

+++ rhum-a-ranger/src/main/java/projetPj/rhum_a_ranger/rhum/RhumService.java
@@ -1,5 +1,6 @@
 package projetPj.rhum_a_ranger.rhum;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -32,6 +33,9 @@ public class RhumService {
     }
 
     public void deleteRhum(Long id) {
+        if (!rhumRepository.existsById(id)) {
+            throw new EntityNotFoundException("Rhum not found with id: " + id);
+        }
         rhumRepository.deleteById(id);
     }
